# docs/swagger.json
{"openapi":"3.0.0","paths":{"/testing/all-data":{"delete":{"operationId":"TestingController_deleteAll","parameters":[],"responses":{"204":{"description":""}},"tags":["Testing"]}},"/sa/users":{"post":{"operationId":"UserSaController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateDtoApi"}}}},"responses":{"201":{"description":""}},"tags":["UserSa"]},"get":{"operationId":"UserSaController_getAllUsers","parameters":[],"responses":{"200":{"description":""}},"tags":["UserSa"]}},"/sa/users/{id}":{"delete":{"operationId":"UserSaController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["UserSa"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginDtoApi"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/password-recovery":{"post":{"operationId":"AuthController_passwordRecovery","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthPasswordRecoveryDtoApi"}}}},"responses":{"204":{"description":""}},"tags":["Auth"]}},"/auth/new-password":{"post":{"operationId":"AuthController_newPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthNewPasswordDtoApi"}}}},"responses":{"204":{"description":""}},"tags":["Auth"]}},"/auth/registration":{"post":{"operationId":"AuthController_registration","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegistrationDtoApi"}}}},"responses":{"204":{"description":""}},"tags":["Auth"]}},"/auth/registration-confirmation":{"post":{"operationId":"AuthController_registrationConfirmation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegistrationConfirmationDtoApi"}}}},"responses":{"204":{"description":""}},"tags":["Auth"]}},"/auth/registration-email-resending":{"post":{"operationId":"AuthController_registrationEmailResend","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegistrationEmailResendingDtoApi"}}}},"responses":{"204":{"description":""}},"tags":["Auth"]}},"/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"204":{"description":""}},"tags":["Auth"]}},"/auth/refresh-token":{"post":{"operationId":"AuthController_refreshToken","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/security/devices":{"get":{"operationId":"SessionController_getAllSessions","parameters":[],"responses":{"200":{"description":""}},"tags":["Session"]},"delete":{"operationId":"SessionController_deleteSessions","parameters":[],"responses":{"204":{"description":""}},"tags":["Session"]}},"/security/devices/{deviceId}":{"delete":{"operationId":"SessionController_deleteSession","parameters":[{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Session"]}},"/blogs":{"get":{"operationId":"BlogsController_getBlogs","parameters":[],"responses":{"200":{"description":""}},"tags":["Blogs"]}},"/blogs/{blogId}/posts":{"get":{"operationId":"BlogsController_getPosts","parameters":[{"name":"blogId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Blogs"]}},"/blogs/{blogId}":{"get":{"operationId":"BlogsController_getBlog","parameters":[{"name":"blogId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Blogs"]}},"/sa/blogs":{"get":{"operationId":"BlogsSaController_getBlogs","parameters":[],"responses":{"200":{"description":""}},"tags":["BlogsSa"]},"post":{"operationId":"BlogsSaController_createBlog","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogCreateDtoApi"}}}},"responses":{"201":{"description":""}},"tags":["BlogsSa"]}},"/sa/blogs/{blogId}":{"put":{"operationId":"BlogsSaController_updateBlog","parameters":[{"name":"blogId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogUpdateDtoApi"}}}},"responses":{"204":{"description":""}},"tags":["BlogsSa"]},"delete":{"operationId":"BlogsSaController_deleteBlog","parameters":[{"name":"blogId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["BlogsSa"]}},"/sa/blogs/{blogId}/posts":{"post":{"operationId":"BlogsSaController_createPostToBlog","parameters":[{"name":"blogId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostToBlogCreateDtoApi"}}}},"responses":{"201":{"description":""}},"tags":["BlogsSa"]},"get":{"operationId":"BlogsSaController_getPosts","parameters":[{"name":"blogId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["BlogsSa"]}},"/sa/blogs/{blogId}/posts/{postId}":{"put":{"operationId":"BlogsSaController_updatePostToBlog","parameters":[{"name":"blogId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["BlogsSa"]},"delete":{"operationId":"BlogsSaController_deletePostToBlog","parameters":[{"name":"blogId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["BlogsSa"]}}},"info":{"title":"INCTAGRAM API","description":"The Inctagram API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserCreateDtoApi":{"type":"object","properties":{}},"AuthLoginDtoApi":{"type":"object","properties":{}},"AuthPasswordRecoveryDtoApi":{"type":"object","properties":{}},"AuthNewPasswordDtoApi":{"type":"object","properties":{}},"AuthRegistrationDtoApi":{"type":"object","properties":{}},"AuthRegistrationConfirmationDtoApi":{"type":"object","properties":{}},"AuthRegistrationEmailResendingDtoApi":{"type":"object","properties":{}},"BlogCreateDtoApi":{"type":"object","properties":{}},"BlogUpdateDtoApi":{"type":"object","properties":{}},"PostToBlogCreateDtoApi":{"type":"object","properties":{}}}}}